core: Change names for refcount clearing operations
These two refcount change operations only apply when erasing all
values from a property (multi-valued or not). Now this is the case,
anyhow rename the enum names so their effect is more clear.

When operating on a specific value for a property, either inserting
or deleting it, the TRACKER_LOG_REF_INC/DEC operations apply.