library: Relax prefix/uri checks in TrackerNamespaceManager
These impositions on namespace URIs being referenced by just one
prefix, or on prefixes not being able to change and point to a
different URI are not anywhere in RDF/turtle definitions.

As this object is used underneath in RDF deserialization to
carry accounting of the RDF data namespaces, there are indeed
ways to trigger these g_error() through RDF parsing, which
is not nice.

Relax these restrictions, as we do not seem to have a reason to
be this strict.

oss-fuzz#444537260