library: Check for closed database on all query/update entrypoints
When we close the database, the thread pools are disposed. This means
queries being attempted after that might result in assertions being hit
when the job cannot be pushed anymore to the thread pools.

Fixes crashes in the test_notifier test, since the TrackerNotifier would
sometimes get caught in an state where the connection was closed right
before this object would attempt to execute a query, with all prior checks
on the cancellable succeeding.