bus: Handle splicing of D-Bus file descriptors
Internally, we have a shortcut for FDs containing data from
a D-Bus UpdateArray operation, this is typically passed from the
TrackerEndpointDBus through the internal TrackerBatch->add_dbus_fd
vmethod, so it can be read directly in a file-based
TrackerSparqlConnection.

This shortcut however does not work (and triggers a crash) if one
tries to create a TrackerEndpointDBus serving a D-Bus based
TrackerSparqlConnection, and batch operations are received through
this endpoint.

This is a convoluted but useful configuration (e.g. the D-Bus buses
are not necessarily the same) so it makes sense to also implement
the TrackerBatch->add_dbus_fd() vmethod for bus-based TrackerBatch
implementations. Yet another shortcut can be taken here by splicing
the FD directly, as opposed to deserializing+serializing it.