core: Add bounds check parsing DOUBLE terminals in RDF/SPARQL
Strenghten the check for the invalid case that we only parse
a single '.' (DOT) character when trying to parse a double by
making sure we are within the string extents, minding that it
may not be NUL-terminated.

While at it, document the function workings a bit better, since
we don't literally follow the BNF production that is documented
in our code (from the SPARQL spec).