core: Fix handling of blank nodes inherited in string properties
The nco:publisher and dc:publisher relationship is relatively strange
since the first has a rdfs:Resource range (nco:Contact to be specific),
but is a subproperty of the latter with a xsd:string range.

Typically, these resources would not be anonymous, and the dc:publisher
property would store a string representation of their IRI. In the case
of setting a nco:publisher property to point to a blank node, the string
inherited in the dc:publisher property would be the string representation
of the blank node "name" (which is stable in TinySPARQL, but that's an
implementation detail), and in serialization to RDF we would detect
this and produce the following output:

  _:bnode123 a nco:Contact .

  <...> a nie:InformationElement;
    nco:publisher _:bnode123 ;
    dc:publisher _:bnode123 .

(i.e. dc:publisher not dealt with as a string) that should result in
a consistent local result when deserialized to another database, with
the dc:publisher string matching the local blank node.

However this is not working as expected, and results in the following
kind of errors during RDF deserialization:

  dc:publisher: Unable to insert multiple values on single valued property (old: urn:bnode:584, new: 584)

When the dc:publisher property is handled both implicitly (as a superproperty
of nco:publisher) and explicitly (as a triple with dc:publisher is specified
in RDF) and an inconsistent conversion to string happens.

Fix the handling of the latter, so the blank node is converted to its string
representation consistently, and no multiple value insertion is deemed to
happen. This is the same string conversion happen during the handling of
property inheritance. This change makes it possible to successfully serialize
and deserialize a full localsearch database locally.